requirements.txt
# Core dependencies
fastapi>=0.95.0,<1.0.0
uvicorn>=0.22.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6,<1.0.0

# Compression libraries
pillow>=10.0.0,<11.0.0
lz4>=4.3.2,<5.0.0
zstandard>=0.20.0,<1.0.0

# Scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0  # Required for sklearn imports
scikit-image>=0.21.0,<1.0.0
imageio>=2.31.0,<3.0.0

# System utilities
psutil>=5.9.0,<6.0.0
requests>=2.31.0,<3.0.0

# Production dependencies
gunicorn>=21.2.0,<22.0.0
httpx>=0.24.0,<1.0.0  # For health checks